http://issues.apache.org/activemq/browse/AMQCPP-30
git-svn-id: https://svn.apache.org/repos/asf/activemq/activemq-cpp/trunk@505777 13f79535-47bb-0310-9956-ffa450edef68

+++ activemq-cpp/src/main/activemq/connector/openwire/marshal/v2/MessageMarshaller.cpp
@@ -111,9 +111,9 @@ int MessageMarshaller::tightMarshal1( OpenWireFormat* wireFormat, DataStructure*
     rc += tightMarshalLong1( wireFormat, info->getTimestamp(), bs );
     rc += tightMarshalString1( info->getType(), bs );
     bs->writeBoolean( info->getContent().size() != 0 );
-    rc += info->getContent().size() == 0 ? 0 : info->getContent().size() + 4;
+    rc += info->getContent().size() == 0 ? 0 : (int)info->getContent().size() + 4;
     bs->writeBoolean( info->getMarshalledProperties().size() != 0 );
-    rc += info->getMarshalledProperties().size() == 0 ? 0 : info->getMarshalledProperties().size() + 4;
+    rc += info->getMarshalledProperties().size() == 0 ? 0 : (int)info->getMarshalledProperties().size() + 4;
     rc += tightMarshalNestedObject1( wireFormat, info->getDataStructure(), bs );
     rc += tightMarshalCachedObject1( wireFormat, info->getTargetConsumerId(), bs );
     bs->writeBoolean( info->isCompressed() );
